cloudinary + create product + deleteproduct , form editproduct stock

+++ backend/src/product/product.service.ts
@@ -7,19 +7,42 @@ import { CreateMultipleProductsDto } from './dto/create-multi.dto';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { Product, ProductDocument } from './schema/product.schema';
+import { CloudinaryService } from 'src/cloudinary/cloudinary.service';
 
 @Injectable()
 export class ProductService {
   constructor(
     @InjectModel(Product.name) private productModel: Model<ProductDocument>,
     @InjectModel(Category.name) private categoryModel: Model<CategoryDocument>,
+    private readonly cloudinaryService: CloudinaryService,
   ) { }
 
   async create(createProductDto: CreateProductDto): Promise<ResponseDto<Product>> {
-      try {
-      const createdProduct = new this.productModel(createProductDto);
+    try {
+      // Tải lên tất cả ảnh từ danh sách base64
+      const uploadedImages = await Promise.all(
+        createProductDto.images.map(async (base64: string) => {
+          const base64Str = base64.split(',')[1]; // Lấy phần dữ liệu base64 (bỏ qua phần prefix)
+          const buffer = Buffer.from(base64Str, 'base64'); // Chuyển đổi base64 thành Buffer
+          const uploadResult = await this.cloudinaryService.uploadMedia(buffer, 'APPLE_STORE', 'image');
+          
+          // Trả về đối tượng với URL và publicId nếu thành công, ngược lại trả về null
+          return uploadResult.success ? { image: uploadResult.data.url, publicId: uploadResult.data.publicId } : null; 
+        })
+      );
+  
+      // Lọc những kết quả hợp lệ (không phải null)
+      const validUploads = uploadedImages.filter(upload => upload !== null) as { image: string; publicId: string }[];
+  
+      // Tạo sản phẩm với danh sách ảnh đã upload
+      const createdProduct = new this.productModel({
+        ...createProductDto,
+        images: validUploads, 
+      });
+  
+      // Lưu sản phẩm vào database
       await createdProduct.save();
-      console.log("🚀 ~ ProductService ~ createdProduct:", createdProduct)
+  
       return {
         success: true,
         message: 'Product created successfully',
@@ -28,11 +51,12 @@ export class ProductService {
     } catch (error) {
       return {
         success: false,
-        message: 'Failed to create product',
+        message: 'Failed to create product: ' + error.message,
         data: null,
       };
     }
   }
+  
 
   async createMultiple(createMultipleProductsDto: CreateMultipleProductsDto): Promise<ResponseDto<Product[]>> {
     try {
@@ -61,7 +85,7 @@ export class ProductService {
     try {
       const skip = (page - 1) * limit;
       const filter: any = {};
-  
+
       // Nếu categoryId không được truyền vào, trả về tất cả sản phẩm
       if (!categoryId) {
         // Thêm bộ lọc giá nếu có
@@ -72,7 +96,7 @@ export class ProductService {
         } else if (maxPrice !== undefined) {
           filter.price = { $lte: maxPrice };
         }
-  
+
         const total = await this.productModel.countDocuments(filter).exec();
         const products = await this.productModel
           .find(filter)
@@ -80,14 +104,14 @@ export class ProductService {
           .limit(limit)
           .skip(skip)
           .exec();
-  
+
         return {
           success: true,
           message: 'Products retrieved successfully',
           data: { products: products, total },
         };
       }
-  
+
       // Tiếp tục xử lý nếu có categoryId
       const currentCategory = await this.categoryModel.findById(categoryId).exec();
       if (!currentCategory) {
@@ -97,21 +121,21 @@ export class ProductService {
           data: { products: [], total: 0 },
         };
       }
-  
+
       // Hàm để tìm tất cả các danh mục con (đệ quy)
       const findAllChildCategories = async (parentId: string): Promise<string[]> => {
         const categoryIds: string[] = [];
         const childCategories = await this.categoryModel.find({ parentCategoryId: parentId }).exec();
-  
+
         for (const child of childCategories) {
           categoryIds.push(child._id.toString()); // Lưu ID của danh mục con
           const grandChildIds = await findAllChildCategories(child._id.toString());
           categoryIds.push(...grandChildIds); // Thêm các danh mục con vào mảng
         }
-  
+
         return categoryIds; // Trả về danh sách tất cả các danh mục con
       };
-  
+
       // Lấy danh sách các danh mục con
       let categoryIds: string[] = [];
       if (currentCategory.parentCategoryId === null) {
@@ -129,12 +153,12 @@ export class ProductService {
           }
         }
       }
-  
+
       // Thêm danh sách danh mục vào filter
       if (categoryIds.length > 0) {
         filter.categoryId = { $in: categoryIds }; // Lọc sản phẩm theo các danh mục
       }
-  
+
       // Thêm bộ lọc giá
       if (minPrice !== undefined && maxPrice !== undefined) {
         filter.price = { $gte: minPrice, $lte: maxPrice };
@@ -143,7 +167,7 @@ export class ProductService {
       } else if (maxPrice !== undefined) {
         filter.price = { $lte: maxPrice };
       }
-  
+
       const total = await this.productModel.countDocuments(filter).exec();
       const products = await this.productModel
         .find(filter)
@@ -152,7 +176,7 @@ export class ProductService {
         .limit(limit)
         .skip(skip)
         .exec();
-  
+
       return {
         success: true,
         message: 'Products retrieved successfully',
@@ -166,7 +190,7 @@ export class ProductService {
       };
     }
   }
-  
+
   async findOne(id: string): Promise<ResponseDto<Product>> {
     try {
       const product = await this.productModel.findById(id).exec();
@@ -214,10 +238,32 @@ export class ProductService {
 
   async remove(id: string): Promise<ResponseDto<Product>> {
     try {
-      const deletedProduct = await this.productModel.findByIdAndDelete(id).exec();
-      if (!deletedProduct) {
+      const productToDelete = await this.productModel.findById(id).exec();
+      if (!productToDelete) {
         throw new NotFoundException(`Product with ID "${id}" not found`);
       }
+  
+      // Lọc ra các publicId hợp lệ (sử dụng _id trong trường hợp này)
+      const validImagePublicIds = productToDelete.images
+        .map(image => image.publicId) // Lấy _id từ từng image
+        .filter(publicId => publicId); // Lọc những publicId không hợp lệ (undefined hoặc null)
+  
+      // Gọi deleteMedia cho từng publicId hợp lệ
+      const deleteMediaPromises = validImagePublicIds.map(async (publicId) => {
+        return this.cloudinaryService.deleteMedia(publicId, 'image');
+      });
+  
+      // Chờ cho tất cả các promise được hoàn thành
+      const deleteMediaResults = await Promise.all(deleteMediaPromises);
+  
+      // Kiểm tra xem tất cả các ảnh đã được xóa thành công
+      const allDeleted = deleteMediaResults.every(result => result.success);
+      if (!allDeleted) {
+        throw new Error('Some media files failed to delete');
+      }
+  
+      // Xóa sản phẩm
+      await this.productModel.findByIdAndDelete(id).exec();
       return {
         success: true,
         message: 'Product deleted successfully',
@@ -229,4 +275,5 @@ export class ProductService {
       };
     }
   }
+  
 }
